flask_amocrm_project/requirements.txt
flasgger==0.9.7.1
Flask==3.0.3
python-dotenv==1.0.0
requests==2.31.0
